refactor: extract editor input handling
